touches fixed

+++ core/src/com/airse/trickynumbers/models/GameButton.java
@@ -22,7 +22,7 @@ public class GameButton extends MyButton {
     private Color textColor;
     private GlyphLayout glyphLayout;
 
-    public GameButton(String text, MyColor buttonColor, int posX, int posY, int width, int height, BitmapFont font){
+    public GameButton(String text, MyColor buttonColor, float posX, float posY, float width, float height, BitmapFont font){
         super(buttonColor, posX, posY, width, height);
 
         this.text = text;
